EPML-4130:change datatype after code review
git-tfs-id: [https://tfs.epmlive.com/tfs/EPMLive/]$/EPMLive2013Release/567-DEV;C6176

+++ EPMLiveWebParts/EPMLiveWebParts/WebPageCode/getgriditems.cs
@@ -3756,7 +3756,7 @@ namespace EPMLiveWebParts
 
                                             if (field.Type == SPFieldType.Calculated)
                                             {
-                                                Int16 itemid = Convert.ToInt16(dr["ItemID"]);
+                                                Int32 itemid = Convert.ToInt32(dr["ItemID"]);
                                                 SPListItem li = list.GetItemById(itemid);
                                                 val = getField(li, field.InternalName, false);
                                                 if (field.Type != SPFieldType.Attachments)
